several code improvements and bug fixes for different execution options

+++ run_simulation.py
@@ -2,13 +2,14 @@ import numpy as np
 import pickle
 from Qiskit_input_graph import draw_graph, calculate_msq
 import modify_graph_objects as mgo
+from witness import witness_fancy, witness_plain, fidelity_est_simple, fidelity_full
 import networkx as nx
 from networkx import Graph
 from qiskit import QuantumCircuit
 from typing import Sequence, Tuple
 import os
 import create_ghz_state_circuit as cgsc
-import calibration_classes as cal
+import simulation_utils as utils
 import copy
 from dotenv import load_dotenv
 from qiskit_ibm_provider import IBMProvider
@@ -43,12 +44,9 @@ else:
     cal_file = args.fname
 
 # get estimator calibration
-est_cal = cal.get_EstimatorCalibration_from_yaml(cal_file)
+est_cal = utils.get_EstimatorCalibration_from_yaml(cal_file)
 # get passmanager calibration
-pm_cal = cal.get_PresetPassManagerCalibration_from_yaml(cal_file)
-
-# load backend from passmanager calibration
-backend = cal.get_backend(pm_cal.backend_str)
+pm_cal = utils.get_PresetPassManagerCalibration_from_yaml(cal_file)
 
 
 # read calibration dictionary
@@ -73,25 +71,6 @@ if result_dir is None:
     print("Could not retrieve a result directory from calibration file. Set it to the current working directory!")
     result_dir = os.getcwd()
 
-graph_dir = cal_dict.get("graph_dir", None)
-graph_fname = cal_dict.get("graph_fname", None)
-if graph_fname is None and graph_dir is None:
-    print("No graph filename and directory were given. Use backend string from passmanager calibration to create graph.")
-    graph = generate_layout_graph(backend)
-else:
-    if graph_fname is None:
-        raise ValueError("Graph directory was given but could not retrieve graph filename! To import from backend directly, leave both empty!")
-    if graph_dir is None:
-        raise ValueError("Graph filename was given but could not retrieve graph directory! To import from backend directly, leave both empty!")
-    graph_file = os.path.join(graph_dir, graph_fname)
-    # load initial graph from pickle
-    graph = None
-    with open(graph_file, "rb") as f:
-        graph = pickle.load(f)
-
-    if not isinstance(graph, Graph):
-        raise ValueError("Loaded graph object is not a networkx graph object!")
-
 # make directory for this simulation run
 result_dir = os.path.join(result_dir, sim_id)
 if os.path.isdir(result_dir):
@@ -99,80 +78,114 @@ if os.path.isdir(result_dir):
 else:
     os.mkdir(result_dir)
 
-# create pass manager from calibration
-pass_manager = cal.get_passmanager(pm_cal)
+
+use_premium_access = cal_dict["use_premium_access"]
+fidelity_witness = cal_dict["fidelity_witness"]
+backend_str = cal_dict["backend_str"]
+noise_model_id = cal_dict["noise_model_id"]
+fname_noise_model = cal_dict["fname_noise_model"]
+fname_noise_model = os.path.join(result_dir, fname_noise_model)
+noise_model_str = cal_dict["noise_model_str"]
+coupling_map_id = cal_dict["coupling_map_id"]
+fname_coupling_map = cal_dict["fname_coupling_map"]
+fname_coupling_map = os.path.join(result_dir, fname_coupling_map)
+coupling_map_str = cal_dict["coupling_map_str"]
+native_basis_gates_str = cal_dict["native_basis_gates_str"]
+# setup backend options
+backend_opt = {"backend_str": backend_str, 
+               "noise_model_id": noise_model_id, 
+               "fname_noise_model": fname_noise_model, 
+               "noise_model_str": noise_model_str,
+               "coupling_map_id": coupling_map_id,
+               "fname_coupling_map": fname_coupling_map,
+               "coupling_map_str": coupling_map_str,
+               "native_basis_gates_str": native_basis_gates_str,
+               "run_locally": cal_dict["run_locally"]}
 
 # load IBM Quantum credentials
-load_dotenv()
-ibmq_api_token = os.environ["TOKEN"]
-ibmq_hub = os.environ["HUB"]
-ibmq_group = os.environ["GROUP"]
-ibmq_project = "multiflavorschwi"
-# ibmq_project = "reservations"
-provider = IBMProvider(token=ibmq_api_token, instance=f"{ibmq_hub}/{ibmq_group}/{ibmq_project}")
-service = QiskitRuntimeService(channel="ibm_quantum", token=ibmq_api_token, instance=f"{ibmq_hub}/{ibmq_group}/{ibmq_project}")
+service = utils.load_ibm_credentials(premium_access=use_premium_access)
 
+# load backend from passmanager calibration
+backend = utils.get_backend(service, backend_opt)
+
+# save noise model if not already done
+noise_model = NoiseModel.from_backend(backend)
+if noise_model:
+    if not noise_model.is_ideal():
+        if not os.path.isfile(fname_noise_model):
+            with open(fname_noise_model, "wb") as f:
+                pickle.dump(noise_model,f)
+
+# save coupling map if not already done
+coupling_map = backend.coupling_map
+if coupling_map:
+    if not os.path.isfile(fname_coupling_map):
+        with open(fname_coupling_map, "wb") as f:
+            pickle.dump(coupling_map,f)
+
+
+subgraph_size = cal_dict.get("subgraph_size", None)
+print("Create layout graph from backend.")
+graph = generate_layout_graph(backend, k=subgraph_size)
+
+# add graph nodes as initial layout to passmanager calibration
+if pm_cal.initial_layout is not None:
+    print("Warning: Initial layout of passmanager calibration is not None but will be overwritten due to algorithm.")
+init_layout = list(range(backend.num_qubits)) # list of all qubits, i.e., a trivial layout i -> i
+pm_cal.initial_layout = init_layout
+
+# create pass manager from calibration
+pass_manager = utils.get_passmanager(backend, pm_cal)
 
 # create the circuit to generate GHZ state
-curr_circ, curr_init_graph, curr_star_graph = cgsc.create_ghz_state_circuit_graph(graph)
+curr_circ, curr_init_graph, curr_star_graph = cgsc.create_ghz_state_circuit_graph(graph, backend.num_qubits)
+#curr_circ, curr_init_graph = cgsc.create_ghz_state_circuit_debug(graph, backend.num_qubits)
 
 # draw graph and save the plot
-fname_graph = f"sim_{sim_id}_layout_graph_backend_{pm_cal.backend_str}.pdf"
+fname_graph = f"sim_{sim_id}_layout_graph_backend_{backend_str}.pdf"
 fname_graph = os.path.join(result_dir, fname_graph)
-mgo.draw_graph(curr_init_graph, title="Layout graph for "+pm_cal.backend_str+ " backend", fname=fname_graph)
+mgo.draw_graph(curr_init_graph, title="Layout graph from "+backend_str+ " backend", fname=fname_graph)
 
-## To-Do integrate funciton to get the observable here, once it's finished!
 # Get fidelity observable
-observalbe = SparsePauliOp(["X"*backend.num_qubits], coeffs=np.asarray([1.0])) # This is just a dummy observable used for debugging
-##
+#observalbe = SparsePauliOp(["X"*backend.num_qubits], coeffs=np.asarray([1.0])) # This is just a dummy observable used for debugging
+if fidelity_witness == "simple":
+    observalbe = fidelity_est_simple(curr_circ.num_qubits, list(graph))
+elif fidelity_witness == "full":
+    observalbe = fidelity_full(curr_circ.num_qubits, list(graph))
+else:
+    raise ValueError("Keyword for fidelity witness {} is unkown".format(fidelity_witness))
 
-# remove all barriers
-curr_circ = RemoveBarriers()(curr_circ)
-## To-Do: add list of graph nodes as initial layout (want to label logical -> physical qubits as i -> i)!
 # transpile circuit with passmanager
-transp_circ = pass_manager.run(curr_circ)
+transp_circ = utils.transpile_circuits(pass_manager, [curr_circ], transpilation_trials=pm_cal.transpilation_trials, remove_barriers=True)[0]
+
 isa_observable = observalbe.apply_layout(transp_circ.layout)
 
+## To-Do: We are running just one circuit! Remove Session environment and run it as a single job.
 with Session(service, backend=backend) as session:
+    ## To-Do: Update Estimator to V2 version
     # create estimator from calibration
-    estimator = cal.get_estimator(est_cal, session=session)
+    estimator = utils.get_estimator(est_cal, mode=session)
+    ##
 
-    # For fake backends to have a true noisy simulation
-    if "fake" in pm_cal.backend_str:
-        # set coupling map for estimator to backend.coupling_map
-        if estimator.options.simulator["coupling_map"] == Unset:
-            estimator.set_options(coupling_map= [list(t) for t in list(backend.coupling_map.get_edges())])
-            # update calibration class
-            est_cal.estimator_options["simulator_options"]["coupling_map"] =  [list(t) for t in list(backend.coupling_map.get_edges())]
-            est_cal.coupling_map_str = pm_cal.backend_str + "_" + datetime.today().strftime('%Y-%m-%d')
-    
-        # set noise model of estimator to backend noise model
-        if estimator.options.simulator["noise_model"] == Unset:
-            estimator.set_options(noise_model = NoiseModel.from_backend(backend))
-            # update calibration class
-            est_cal.estimator_options["simulator_options"]["noise_model"] = NoiseModel.from_backend(backend)
-            est_cal.noise_model_str = pm_cal.backend_str + "_" + datetime.today().strftime('%Y-%m-%d')
-            
-    
-        if estimator.options.simulator["basis_gates"] == Unset:
-            estimator.set_options(basis_gates = NoiseModel.from_backend(backend).basis_gates)
-            # update calibration class
-            est_cal.estimator_options["simulator_options"]["basis_gates"] = NoiseModel.from_backend(backend).basis_gates
-            est_cal.basis_gates_str = pm_cal.backend_str + "_" + datetime.today().strftime('%Y-%m-%d')
-    
     # run circuit and observable on backend
     job = estimator.run(transp_circ, isa_observable)
     est_result = job.result()
 
     fidelity = est_result.values[0]
+##
 
 # save the result and all calibration data, graph, transpiled circuit, etc. in the result dir
-## To-Do: passmanager calibration data, that is None must be converted to string "None" before writing to csv file
 # save to csv file
 header_est, data_est = est_cal.get_filevector()
 header_pm, data_pm = pm_cal.get_filevector()
-csv_header = ["fidelity"] + header_est + header_pm
-csv_data = [fidelity] + data_est + data_pm
+csv_header = ["fidelity"] 
+for k in backend_opt.keys():
+    csv_header = csv_header + [k]
+csv_header = csv_header + header_est + header_pm
+csv_data = [fidelity] 
+for v in backend_opt.values():
+    csv_data = csv_data + [v]
+csv_data = csv_data + data_est + data_pm
 fname_csv = f"sim_{sim_id}_results.csv"
 fname_csv = os.path.join(result_dir, fname_csv)
 
@@ -182,39 +195,44 @@ with open(fname_csv, "w", newline="") as csvfile:
     spamwriter.writerow(csv_data)
 ##
 
-# save calibration as yaml files
-## To-Do: Save both calibration data in one file (similar as the input calibration file)
-fname_est_cal = f"sim_{sim_id}_estimator_calibration.yaml"
-fname_est_cal = os.path.join(result_dir, fname_est_cal)
-est_cal.to_yaml(fname_est_cal)
+# save calibration as yaml file
+fname_cal = f"sim_{sim_id}_calibration.yaml"
+fname_cal = os.path.join(result_dir, fname_cal)
+# check if file already exists
+if os.path.isfile(fname_cal):
+    raise ValueError("file {} does already exist!".format(fname_cal))
+# prepare calibration classes for yaml
+est_cal_dict = est_cal.to_dict()
+pm_cal_dict = pm_cal.to_dict()
+cal_dict["estimator_calibration"] = est_cal_dict
+cal_dict["passmanager_calibration"] = pm_cal_dict
+# dump calibration dictionary into yaml file
+with open(fname_cal, "w") as f:
+    yaml.dump(cal_dict, f)
 
-fname_pm_cal = f"sim_{sim_id}_passmanager_calibration.yaml"
-fname_pm_cal = os.path.join(result_dir, fname_pm_cal)
-pm_cal.to_yaml(fname_pm_cal)
-##
 
 # pickle intial layout graph
-fname_graph = f"sim_{sim_id}_layout_graph_backend_{pm_cal.backend_str}.pickle"
+fname_graph = f"sim_{sim_id}_layout_graph_backend_{backend_str}.pickle"
 fname_graph = os.path.join(result_dir, fname_graph)
 with open(fname_graph, "wb") as f:
     pickle.dump(curr_init_graph,f)
 
 # pickle initial circuit and save plot
-fname_circ = f"sim_{sim_id}_circuit_backend_{pm_cal.backend_str}.pickle"
+fname_circ = f"sim_{sim_id}_circuit_backend_{backend_str}.pickle"
 fname_circ = os.path.join(result_dir, fname_circ)
 with open(fname_circ, "wb") as f:
     pickle.dump(curr_circ, f)
-fname_circ = f"sim_{sim_id}_circuit_backend_{pm_cal.backend_str}.pdf"
+fname_circ = f"sim_{sim_id}_circuit_backend_{backend_str}.pdf"
 fname_circ = os.path.join(result_dir, fname_circ)
 curr_circ.draw("mpl", filename=fname_circ)
 
 
 # pickle transpiled circuit and save plot
-fname_circ_transp = f"sim_{sim_id}_circuit_transpiled_backend_{pm_cal.backend_str}.pickle"
+fname_circ_transp = f"sim_{sim_id}_circuit_transpiled_backend_{backend_str}.pickle"
 fname_circ_transp = os.path.join(result_dir, fname_circ_transp)
 with open(fname_circ_transp, "wb") as f:
     pickle.dump(transp_circ, f)
-fname_circ_transp = f"sim_{sim_id}_circuit_transpiled_backend_{pm_cal.backend_str}.pdf"
+fname_circ_transp = f"sim_{sim_id}_circuit_transpiled_backend_{backend_str}.pdf"
 fname_circ_transp = os.path.join(result_dir, fname_circ_transp)
 transp_circ.draw("mpl", filename=fname_circ_transp)
